+++ src/pages/games/SnackGame/AppleGamePage.tsx
