+++ src/pages/games/SnackGame/RankingPage.tsx
@@ -4,8 +4,8 @@ import { Helmet } from 'react-helmet-async';
 import Dropdown, { DropDownOptionType } from '@components/DropDown/DropDown';
 import Footer from '@components/Footer/Footer';
 import Spacing from '@components/Spacing/Spacing';
-import AppleGameHeader from '@pages/games/AppleGame/components/AppleGameHeader';
-import RankingSection from '@pages/games/AppleGame/components/RankingSection';
+import AppleGameHeader from '@pages/games/SnackGame/components/AppleGameHeader';
+import RankingSection from '@pages/games/SnackGame/components/RankingSection';
 
 const RankingPage = () => {
   const [selectedSeason, setSelectedSeason] = useState<number>(1);
