+++ src/pages/games/SnackGame/components/AppleGameHeader.tsx
