+++ src/pages/games/SnackGame/components/ExpBarChart.tsx
