+++ src/pages/games/SnackGame/components/GameResult.tsx
