+++ src/pages/games/SnackGame/components/OtherRanking.tsx
