+++ src/pages/games/SnackGame/components/TopRanking.tsx
