+++ src/pages/games/SnackGame/components/UserRanking.tsx
