+++ src/pages/games/SnackGame/game/game.type.ts
