+++ src/pages/games/SnackGame/game/model/appleGame/apple.ts
