+++ src/pages/games/SnackGame/game/model/appleGame/goldenApple.ts
