+++ src/pages/games/SnackGame/game/model/appleGame/plainApple.ts
