+++ src/pages/games/SnackGame/game/model/common/click.ts
