+++ src/pages/games/SnackGame/game/model/common/drag.ts
