+++ src/pages/games/SnackGame/game/model/common/particle.ts
