+++ src/pages/games/SnackGame/game/model/snackGame/goldSnack.ts
