+++ src/pages/games/SnackGame/game/model/snackGame/plainSnack.ts
