+++ src/pages/games/SnackGame/game/model/snackGame/snack.ts
