+++ src/pages/games/SnackGame/game/model/snackGame/snackGame.ts
