+++ src/pages/games/SnackGame/game/model/snackGame/snackGameB.ts
