+++ src/pages/games/SnackGame/game/model/snackGame/snackGameC.ts
