+++ src/pages/games/SnackGame/game/model/snackGame/snackGameD.ts
