+++ src/pages/games/SnackGame/game/view/SnackGameHUD.tsx
