+++ src/pages/games/SnackGame/game/view/appleGame/DefaultMode.tsx
@@ -2,16 +2,16 @@ import { useEffect, useState } from 'react';
 
 import QueryBoundary from '@components/base/QueryBoundary';
 import retryError from '@components/Error/RetryError';
-import GameResult from '@pages/games/AppleGame/components/GameResult';
+import GameResult from '@pages/games/SnackGame/components/GameResult';
 import {
   goldModAppleType,
   scoredAppleRectType,
-} from '@pages/games/AppleGame/game/game.type';
-import Apple from '@pages/games/AppleGame/game/model/appleGame/apple';
-import { AppleGame } from '@pages/games/AppleGame/game/model/appleGame/appleGame';
-import { GoldenApple } from '@pages/games/AppleGame/game/model/appleGame/goldenApple';
-import PlainApple from '@pages/games/AppleGame/game/model/appleGame/plainApple';
-import SnackGameHUD from '@pages/games/AppleGame/game/view/SnackGameHUD';
+} from '@pages/games/SnackGame/game/game.type';
+import Apple from '@pages/games/SnackGame/game/model/appleGame/apple';
+import { AppleGame } from '@pages/games/SnackGame/game/model/appleGame/appleGame';
+import { GoldenApple } from '@pages/games/SnackGame/game/model/appleGame/goldenApple';
+import PlainApple from '@pages/games/SnackGame/game/model/appleGame/plainApple';
+import SnackGameHUD from '@pages/games/SnackGame/game/view/SnackGameHUD';
 
 import {
   useAppleGameRefresh,
